Move layout components to their own module and rename paths
