--- client/components/layout/index.js
@@ -1,4 +1,0 @@
-export Block from './Block'
-export Container from './Container'
-export Content from './Content'
-export Frame from './Frame'
