+++ client/modules/layout/components/Container/__tests__/Container-spec.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import expectReactShallow from 'expect-react-shallow'
 
-import Base from '../../internal/Base'
+import Base from '../../Base'
 import Container from '..'
 
 describe('Container', () => {
