+++ client/modules/layout/components/Content/__tests__/Content-spec.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import expectReactShallow from 'expect-react-shallow'
 
-import Base from '../../internal/Base'
+import Base from '../../Base'
 import Content from '..'
 
 describe('Content', () => {
