+++ client/modules/layout/index.js
@@ -1,0 +1,4 @@
+export Block from './components/Block'
+export Container from './components/Container'
+export Content from './components/Content'
+export Frame from './components/Frame'
